Rust 1.95

+++ wasm-lib/src/review_functions.rs
@@ -37,7 +37,7 @@ pub fn review_card(
     if let Some(last_session) = card.reviews.last()
         && let Some(last_date) = parse_datetime_flexible(&last_session.date)
     {
-        let elapsed_days = (now - last_date).num_days().max(0) as i64;
+        let elapsed_days = (now - last_date).num_days().max(0);
         fsrs_card.elapsed_days = elapsed_days;
     }
