+++ wasm-lib/src/state_functions.rs
@@ -39,7 +39,7 @@ pub fn compute_current_state(
     if let Some(last_session) = last_session
         && let Some(last_date) = parse_datetime_flexible(&last_session.date)
     {
-        let elapsed_days = (now - last_date).num_days().max(0) as i64;
+        let elapsed_days = (now - last_date).num_days().max(0);
         fsrs_card.elapsed_days = elapsed_days;
         fsrs_card.last_review = last_date;
